[feat] Add viewOfPost by ctg
카테고리에 맞는 게시물들을 출력하여 화면에 보여주는 메소드 작성, 별도의 클래스로 분리.
카테고리 리스트, 로고 등을 보여주던 우측 에셋을 구현하는 메소드를 별도의 클래스로 분리
